Move converted external tests to local.rs
